Loop environment: created LoopEnvironmentUser.hpp
