+++ src/core/loop_environment/src/CMakeLists.txt
@@ -1,6 +1,7 @@
 # Sources
 set(Srcs 
   LoopEnvironment.cpp
+  LoopEnvironmentUser.cpp
   EnvBuilder.cpp
 )
