+++ src/core/loop_environment/src/EnvBuilder.cpp
@@ -14,92 +14,6 @@
 using namespace llvm;
 using namespace llvm::noelle;
 
-EnvUserBuilder::EnvUserBuilder ()
-  : envIndexToPtr{}, liveInInds{}, liveOutInds{} {
-  envIndexToPtr.clear();
-  liveInInds.clear();
-  liveOutInds.clear();
-}
-
-EnvUserBuilder::~EnvUserBuilder () {
-}
-
-void EnvUserBuilder::createEnvPtr (
-  IRBuilder<> builder,
-  int envIndex,
-  Type *type
-) {
-
-  /*
-   * Check the inputs.
-   */
-  if (!this->envArray) {
-    errs() << "A reference to the environment array has not been set for this user!\n";
-    abort();
-  }
-
-  /*
-   * Create the zero integer constant.
-   */
-  auto int64 = IntegerType::get(builder.getContext(), 64);
-  auto zeroV = cast<Value>(ConstantInt::get(int64, 0));
-
-  /*
-   * Compute how many values can fit in a cache line.
-   */
-  auto valuesInCacheLine = Architecture::getCacheLineBytes() / sizeof(int64_t);
-
-  auto envIndV = cast<Value>(ConstantInt::get(int64, envIndex * valuesInCacheLine));
-
-  auto envGEP = builder.CreateInBoundsGEP(
-    this->envArray,
-    ArrayRef<Value*>({ zeroV, envIndV })
-  );
-  auto envPtr = builder.CreateBitCast(envGEP, PointerType::getUnqual(type));
-
-  this->envIndexToPtr[envIndex] = cast<Instruction>(envPtr);
-}
-
-void EnvUserBuilder::createReducableEnvPtr (
-  IRBuilder<> builder,
-  int envIndex,
-  Type *type,
-  int reducerCount,
-  Value *reducerIndV
-) {
-  if (!this->envArray) {
-    errs() << "A reference to the environment array has not been set for this user!\n";
-    abort();
-  }
-
-  /*
-   * Compute how many values can fit in a cache line.
-   */
-  auto valuesInCacheLine = Architecture::getCacheLineBytes() / sizeof(int64_t);
-
-  auto int8 = IntegerType::get(builder.getContext(), 8);
-  auto ptrTy_int8 = PointerType::getUnqual(int8);
-  auto int64 = IntegerType::get(builder.getContext(), 64);
-  auto zeroV = cast<Value>(ConstantInt::get(int64, 0));
-  auto envIndV = cast<Value>(ConstantInt::get(int64, envIndex * valuesInCacheLine));
-
-  auto envReduceGEP = builder.CreateInBoundsGEP(
-    this->envArray,
-    ArrayRef<Value*>({ zeroV, envIndV })
-  );
-  auto arrPtr = PointerType::getUnqual(ArrayType::get(int64, reducerCount * valuesInCacheLine));
-  auto envReducePtr = builder.CreateBitCast(envReduceGEP, PointerType::getUnqual(arrPtr));
-
-  auto reduceIndAlignedV = builder.CreateMul(reducerIndV, ConstantInt::get(int64, valuesInCacheLine));
-  auto envGEP = builder.CreateInBoundsGEP(
-    builder.CreateLoad(envReducePtr),
-    ArrayRef<Value*>({ zeroV, reduceIndAlignedV })
-  );
-  auto envPtr = builder.CreateBitCast(envGEP, PointerType::getUnqual(type));
-
-  this->envIndexToPtr[envIndex] = cast<Instruction>(envPtr);
-}
-
 EnvBuilder::EnvBuilder (LLVMContext &cxt)
   : CXT{cxt}, envTypes{}, envUsers{},
     envIndexToVar{}, envIndexToReducableVar{}, envIndexToVectorOfReducableVar{},
